backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.3",
    "langchain-google-genai>=4.1.3",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "supabase>=2.27.1",
    "uvicorn>=0.40.0",
    "websockets>=15.0.1",
]
